requirements.txt
fastapi
uvicorn[standard]
jinja2
websockets
google-generativeai
python-dotenv
pandas
matplotlib
seaborn
weasyprint
scikit-learn
numpy
openpyxl
python-docx